// frontend/src/generated-schema.ts
export const SYSTEM_TABLE_NAMES = {
    SYSTEM_ACTION: '_System_Action',
    SYSTEM_APP: '_System_App',
    SYSTEM_APPROVALPROCESS: '_System_ApprovalProcess',
    SYSTEM_APPROVALWORKITEM: '_System_ApprovalWorkItem',
    SYSTEM_AUDITLOG: '_System_AuditLog',
    SYSTEM_AUTONUMBER: '_System_AutoNumber',
    SYSTEM_COMMENT: '_System_Comment',
    SYSTEM_CONFIG: '_System_Config',
    SYSTEM_DASHBOARD: '_System_Dashboard',
    SYSTEM_EMAILTEMPLATE: '_System_EmailTemplate',
    SYSTEM_FEEDITEM: '_System_FeedItem',
    SYSTEM_FIELD: '_System_Field',
    SYSTEM_FIELDDEPENDENCY: '_System_FieldDependency',
    SYSTEM_FIELDPERMS: '_System_FieldPerms',
    SYSTEM_FILE: '_System_File',
    SYSTEM_FLOW: '_System_Flow',
    SYSTEM_FLOWINSTANCE: '_System_FlowInstance',
    SYSTEM_FLOWSTEP: '_System_FlowStep',
    SYSTEM_GROUP: '_System_Group',
    SYSTEM_GROUPMEMBER: '_System_GroupMember',
    SYSTEM_LAYOUT: '_System_Layout',
    SYSTEM_LISTVIEW: '_System_ListView',
    SYSTEM_LOG: '_System_Log',
    SYSTEM_NOTIFICATION: '_System_Notification',
    SYSTEM_OBJECT: '_System_Object',
    SYSTEM_OBJECTPERMS: '_System_ObjectPerms',
    SYSTEM_OUTBOXEVENT: '_System_OutboxEvent',
    SYSTEM_PERMISSIONSET: '_System_PermissionSet',
    SYSTEM_PERMISSIONSETASSIGNMENT: '_System_PermissionSetAssignment',
    SYSTEM_PROFILE: '_System_Profile',
    SYSTEM_PROFILELAYOUT: '_System_ProfileLayout',
    SYSTEM_RECENT: '_System_Recent',
    SYSTEM_RECORDSHARE: '_System_RecordShare',
    SYSTEM_RECORDTYPE: '_System_RecordType',
    SYSTEM_RECYCLEBIN: '_System_RecycleBin',
    SYSTEM_RELATIONSHIP: '_System_Relationship',
    SYSTEM_ROLE: '_System_Role',
    SYSTEM_SESSION: '_System_Session',
    SYSTEM_SETUPPAGE: '_System_SetupPage',
    SYSTEM_SHARINGRULE: '_System_SharingRule',
    SYSTEM_TABLE: '_System_Table',
    SYSTEM_TEAMMEMBER: '_System_TeamMember',
    SYSTEM_THEME: '_System_Theme',
    SYSTEM_UICOMPONENT: '_System_UIComponent',
    SYSTEM_USER: '_System_User',
    SYSTEM_VALIDATION: '_System_Validation',
    SYSTEM_WEBHOOK: '_System_Webhook',
} as const;
// ...
export const FIELDS_SYSTEM_FEEDITEM = {
    BODY: 'body',
    CREATED_BY_ID: 'created_by_id',
    CREATED_DATE: 'created_date',
    ID: 'id',
    LAST_MODIFIED_DATE: 'last_modified_date',
    PARENT_ID: 'parent_id',
    TYPE: 'type',
} as const;